Remove global sass->css, modular structure
