+++ src/ui/components/app.js
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact';
 import { Router } from 'preact-router';
 
-import Header from './Header';
+import Header from './header';
 
 // Code-splitting is automated for routes
 import Home from '../routes/home';
