+++ src/ui/components/header/style.scss
@@ -1,0 +1,6 @@
+@import "../../sass/brand.scss";
+@import "../../sass/bulma.scss";
+
+nav {
+	font-family: 'Red Rose', 'Nunito', sans-serif;
+}
