+++ src/ui/index.js
@@ -1,4 +1,3 @@
-import '../../dist/all.css';
 import App from './components/app';
 
 export default App;
