--- src/ui/sass/all.scss
@@ -1,6 +1,0 @@
-@charset "utf-8";
-
-@import './base-helper.scss';
-@import './header.scss';
-
-@import "./bulma.scss";
