+++ src/ui/sass/header.scss
@@ -1,4 +1,4 @@
-// @import url('https://fonts.googleapis.com/css2?family=Red+Rose:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Red+Rose:wght@400;700&display=swap');
 
 nav {
 	font-family: 'Red Rose', 'Nunito', sans-serif;
